Move old ssl logic for connectors into LegacySslProvider
- Inject LegacySslContextFactoryProvider if not using ssl/ssl-provider syntax
- Update JDisc unit tests to use new connector config
