+++ config-model/src/main/java/com/yahoo/vespa/model/container/http/ssl/LegacySslProvider.java
@@ -5,24 +5,24 @@ import com.yahoo.component.ComponentId;
 import com.yahoo.container.bundle.BundleInstantiationSpecification;
 import com.yahoo.jdisc.http.ConnectorConfig;
 import com.yahoo.jdisc.http.ssl.SslContextFactoryProvider;
-import com.yahoo.jdisc.http.ssl.ThrowingSslContextFactoryProvider;
+import com.yahoo.jdisc.http.ssl.LegacySslContextFactoryProvider;
 import com.yahoo.osgi.provider.model.ComponentModel;
 import com.yahoo.vespa.model.container.component.SimpleComponent;
 
 import static com.yahoo.component.ComponentSpecification.fromString;
 
 /**
- * Provides a dummy implementation of {@link SslContextFactoryProvider} to be injected into non-ssl connectors
+ * Provides a legacy implementation of {@link SslContextFactoryProvider} to be injected into non-ssl connectors and connectors using legacy ssl config override
  *
  * @author bjorncs
  */
-public class DummySslProvider extends SimpleComponent implements ConnectorConfig.Producer {
+public class LegacySslProvider extends SimpleComponent implements ConnectorConfig.Producer {
 
-    public static final String COMPONENT_ID_PREFIX = "dummy-ssl-provider@";
-    public static final String COMPONENT_CLASS = ThrowingSslContextFactoryProvider.class.getName();
+    public static final String COMPONENT_ID_PREFIX = "legacy-ssl-provider@";
+    public static final String COMPONENT_CLASS = LegacySslContextFactoryProvider.class.getName();
     public static final String COMPONENT_BUNDLE = "jdisc_http_service";
 
-    public DummySslProvider(String serverName) {
+    public LegacySslProvider(String serverName) {
         super(new ComponentModel(
                 new BundleInstantiationSpecification(new ComponentId(COMPONENT_ID_PREFIX + serverName),
                                                      fromString(COMPONENT_CLASS),
@@ -30,5 +30,7 @@ public class DummySslProvider extends SimpleComponent implements ConnectorConfig
     }
 
     @Override
-    public void getConfig(ConnectorConfig.Builder builder) {}
+    public void getConfig(ConnectorConfig.Builder builder) {
+
+    }
 }
