+++ jdisc_http_service/src/main/java/com/yahoo/jdisc/http/ssl/DefaultSslContextFactoryProvider.java
@@ -31,22 +31,17 @@ import java.util.logging.Logger;
  */
 public class DefaultSslContextFactoryProvider implements SslContextFactoryProvider {
 
-    private static final Logger log = Logger.getLogger(DefaultSslContextFactoryProvider.class.getName());
-
     private final ConnectorConfig connectorConfig;
-    @SuppressWarnings("deprecation")
-    private final com.yahoo.jdisc.http.SecretStore secretStore;
 
-    public DefaultSslContextFactoryProvider(ConnectorConfig connectorConfig,
-                                            @SuppressWarnings("deprecation") com.yahoo.jdisc.http.SecretStore secretStore) {
+    public DefaultSslContextFactoryProvider(ConnectorConfig connectorConfig) {
         validateConfig(connectorConfig.ssl());
         this.connectorConfig = connectorConfig;
-        this.secretStore = secretStore;
     }
 
     @Override
     public SslContextFactory getInstance(String containerId, int port) {
         ConnectorConfig.Ssl sslConfig = connectorConfig.ssl();
+        if (!sslConfig.enabled()) throw new IllegalStateException();
         SslContextFactory factory = new JDiscSslContextFactory();
 
         switch (sslConfig.clientAuth()) {
@@ -67,116 +62,24 @@ public class DefaultSslContextFactoryProvider implements SslContextFactoryProvid
         factory.setExcludeCipherSuites(excludedCiphersWithoutTlsRsaExclusion);
 
         // Check if using new ssl syntax from services.xml
-        if (!sslConfig.privateKeyFile().isEmpty()) {
-            factory.setKeyStore(createKeystore(sslConfig));
-            if (!sslConfig.caCertificateFile().isEmpty()) {
-                factory.setTrustStore(createTruststore(sslConfig));
-            }
-            factory.setProtocol("TLS");
-        } else { // TODO Vespa 7: Remove support for deprecated ssl connector config
-            configureUsingDeprecatedConnectorConfig(sslConfig, factory);
+        factory.setKeyStore(createKeystore(sslConfig));
+        factory.setKeyStorePassword("");
+        if (!sslConfig.caCertificateFile().isEmpty()) {
+            factory.setTrustStore(createTruststore(sslConfig));
         }
+        factory.setProtocol("TLS");
         return factory;
-    }
-
-    private void configureUsingDeprecatedConnectorConfig(ConnectorConfig.Ssl sslConfig, SslContextFactory factory) {
-        switch (sslConfig.keyStoreType()) {
-            case JKS:
-                factory.setKeyStorePath(sslConfig.keyStorePath());
-                factory.setKeyStoreType("JKS");
-                factory.setKeyStorePassword(secretStore.getSecret(sslConfig.keyDbKey()));
-                break;
-            case PEM:
-                factory.setKeyStorePath(sslConfig.keyStorePath());
-                factory.setKeyStore(createPemKeyStore(sslConfig.pemKeyStore()));
-                break;
-        }
-
-        if (!sslConfig.trustStorePath().isEmpty()) {
-            factory.setTrustStorePath(sslConfig.trustStorePath());
-            factory.setTrustStoreType(sslConfig.trustStoreType().toString());
-            if (sslConfig.useTrustStorePassword()) {
-                factory.setTrustStorePassword(secretStore.getSecret(sslConfig.keyDbKey()));
-            }
-        }
-
-        if (!sslConfig.prng().isEmpty()) {
-            factory.setSecureRandomAlgorithm(sslConfig.prng());
-        }
-
-        setStringArrayParameter(
-                factory, sslConfig.excludeProtocol(), ConnectorConfig.Ssl.ExcludeProtocol::name, SslContextFactory::setExcludeProtocols);
-        setStringArrayParameter(
-                factory, sslConfig.includeProtocol(), ConnectorConfig.Ssl.IncludeProtocol::name, SslContextFactory::setIncludeProtocols);
-        setStringArrayParameter(
-                factory, sslConfig.excludeCipherSuite(), ConnectorConfig.Ssl.ExcludeCipherSuite::name, SslContextFactory::setExcludeCipherSuites);
-        setStringArrayParameter(
-                factory, sslConfig.includeCipherSuite(), ConnectorConfig.Ssl.IncludeCipherSuite::name, SslContextFactory::setIncludeCipherSuites);
-
-        factory.setKeyManagerFactoryAlgorithm(sslConfig.sslKeyManagerFactoryAlgorithm());
-        factory.setProtocol(sslConfig.protocol());
     }
 
     private static void validateConfig(ConnectorConfig.Ssl config) {
         if (!config.enabled()) return;
-        if (!config.privateKeyFile().isEmpty()) {
-            if (config.certificateFile().isEmpty()) {
-                throw new IllegalArgumentException("Missing certificate file.");
-            }
-        } else {
-            validateConfigUsingDeprecatedConnectorConfig(config);
+        if (config.certificateFile().isEmpty()) {
+            throw new IllegalArgumentException("Missing certificate file.");
         }
-    }
+        if (config.privateKeyFile().isEmpty()) {
+            throw new IllegalArgumentException("Missing private key file.");
+        }
 
-    private static void validateConfigUsingDeprecatedConnectorConfig(ConnectorConfig.Ssl config) {
-        switch (config.keyStoreType()) {
-            case JKS:
-                validateJksConfig(config);
-                break;
-            case PEM:
-                validatePemConfig(config);
-                break;
-        }
-        if (!config.trustStorePath().isEmpty() && config.useTrustStorePassword() && config.keyDbKey().isEmpty()) {
-            throw new IllegalArgumentException("Missing password for JKS truststore");
-        }
-    }
-
-    private static void validateJksConfig(ConnectorConfig.Ssl ssl) {
-        if (!ssl.pemKeyStore().keyPath().isEmpty() || ! ssl.pemKeyStore().certificatePath().isEmpty()) {
-            throw new IllegalArgumentException("pemKeyStore attributes can not be set when keyStoreType is JKS.");
-        }
-        if (ssl.keyDbKey().isEmpty()) {
-            throw new IllegalArgumentException("Missing password for JKS keystore");
-        }
-    }
-
-    private static void validatePemConfig(ConnectorConfig.Ssl ssl) {
-        if (! ssl.keyStorePath().isEmpty()) {
-            throw new IllegalArgumentException("keyStorePath can not be set when keyStoreType is PEM");
-        }
-        if (!ssl.keyDbKey().isEmpty()) {
-            // TODO Make an error once there are separate passwords for truststore and keystore
-            log.warning("Encrypted PEM key stores are not supported. Password is only applied to truststore");
-        }
-        if (ssl.pemKeyStore().certificatePath().isEmpty()) {
-            throw new IllegalArgumentException("Missing certificate path.");
-        }
-        if (ssl.pemKeyStore().keyPath().isEmpty()) {
-            throw new IllegalArgumentException("Missing key path.");
-        }
-    }
-
-    private static KeyStore createPemKeyStore(ConnectorConfig.Ssl.PemKeyStore pemKeyStore) {
-        try {
-            Path certificatePath = Paths.get(pemKeyStore.certificatePath());
-            Path keyPath = Paths.get(pemKeyStore.keyPath());
-            return new PemSslKeyStore(certificatePath, keyPath).loadJavaKeyStore();
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        } catch (Exception e) {
-            throw new RuntimeException("Failed setting up key store for " + pemKeyStore.keyPath() + ", " + pemKeyStore.certificatePath(), e);
-        }
     }
 
     private static KeyStore createTruststore(ConnectorConfig.Ssl sslConfig) {
@@ -202,13 +105,4 @@ public class DefaultSslContextFactoryProvider implements SslContextFactoryProvid
         }
     }
 
-    private static <T extends InnerNode> void setStringArrayParameter(SslContextFactory sslContextFactory,
-                                                                      List<T> configValues,
-                                                                      Function<T, String> nameProperty,
-                                                                      BiConsumer<SslContextFactory, String[]> setter) {
-        if (!configValues.isEmpty()) {
-            String[] nameArray = configValues.stream().map(nameProperty).toArray(String[]::new);
-            setter.accept(sslContextFactory, nameArray);
-        }
-    }
 }
